requirements.txt
google-adk>=0.1.0
google-genai>=0.3.0 
python-dotenv